+++ comics/comics/undeclaredmajor.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -7,22 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.undeclaredcomics.com/'
     start_date = '2011-08-09'
+    end_date = '2012-09-11'
+    active = False
     rights = 'Belal'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 180
-    schedule = 'Tu'
-    time_zone = 'US/Eastern'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://www.undeclaredcomics.com/feed/')
-        for entry in feed.for_date(pub_date):
-            if 'Comics' not in entry.tags:
-                continue
-            url = entry.summary.src('img.comicthumbnail')
-            if not url:
-                continue
-            url = url.replace('comics-rss', 'comics')
-            title = entry.title
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published